Feature: * Check if the domain event being replayed actually belongs to the entity/aggregate.
