+++ src/EricksonReyes/DomainDrivenDesign/Common/Exception/DomainEventOwnershipException.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace EricksonReyes\DomainDrivenDesign\Common\Exception;
+
+
+use BadMethodCallException;
+
+final class DomainEventOwnershipException extends BadMethodCallException
+{
+
+}
