+++ src/EricksonReyes/DomainDrivenDesign/Common/Exception/PermissionDeniedException.php
@@ -1,10 +1,4 @@
 <?php
-/**
- * Created by PhpStorm.
- * User: erickson
- * Date: 04/10/2018
- * Time: 6:25 PM
- */
 
 namespace EricksonReyes\DomainDrivenDesign\Common\Exception;
